chore(data-workflows): Updated components structure
