+++ src/components/data-workflows/gcs-to-gbq/GcsToGbqConfs.vue
@@ -158,9 +158,9 @@
 import { mapState } from 'vuex';
 import { mapGetters } from 'vuex';
 import VueJsonPretty from 'vue-json-pretty';
-import store from '@/store/index';
-import DataManagementFilters from './widgets/filters/DataManagementFilters';
-import ActivatedStatusChip from './widgets/datatablewidgets/ActivatedStatusChip.vue';
+import store from '@/store';
+import DataManagementFilters from '../../widgets/filters/DataManagementFilters';
+import ActivatedStatusChip from '../../widgets/datatablewidgets/ActivatedStatusChip.vue';
 import ConfsComponent from '@/mixins/confsComponent.js';
 
 export default {
