+++ src/components/data-workflows/gcs-to-gcs/GcsToGcsConfs.vue
@@ -155,11 +155,11 @@
 import { mapState } from 'vuex';
 import { mapGetters } from 'vuex';
 import VueJsonPretty from 'vue-json-pretty';
-import store from '@/store/index';
+import store from '@/store';
 import _ from 'lodash';
-import ActivatedStatusChip from './widgets/datatablewidgets/ActivatedStatusChip.vue';
+import ActivatedStatusChip from '../../widgets/datatablewidgets/ActivatedStatusChip.vue';
 import ConfsComponent from '@/mixins/confsComponent.js';
-import DataManagementFilters from './widgets/filters/DataManagementFilters';
+import DataManagementFilters from '../../widgets/filters/DataManagementFilters';
 
 export default {
 	mixins: [ConfsComponent],
