+++ src/components/data-workflows/gcs-to-gcs/GcsToGcsRun.vue
@@ -17,20 +17,12 @@
 					</v-tab>
 					<v-tab-item value="rundetails">
 						<v-card v-if="run !== undefined">
-							<StorageToStorageRunView
-								:run="run"
-								:runId="run.triggering_file"
-								:activeHeader="true"
-							></StorageToStorageRunView>
+							<GcsToGcsRunView :run="run" :runId="run.gcs_triggering_file" :activeHeader="true"></GcsToGcsRunView>
 						</v-card>
 					</v-tab-item>
 					<v-tab-item value="runconfiguration">
 						<v-card v-if="run !== undefined">
-							<StorageToStorageConfView
-								:conf="run.configuration_context"
-								:configurationId="run.configuration_id"
-								:activeHeader="false"
-							></StorageToStorageConfView>
+							<GcsToGcsConfView :conf="run.configuration_context" :activeHeader="false"></GcsToGcsConfView>
 						</v-card>
 					</v-tab-item>
 					<v-tab-item value="fulljson">
@@ -51,15 +43,15 @@
 
 <script>
 import { mapState } from 'vuex';
-import StorageToStorageConfView from './widgets/configurations/StorageToStorageConfView';
-import StorageToStorageRunView from './widgets/runs/StorageToStorageRunView';
-import store from '@/store/index';
+import GcsToGcsConfView from '../../widgets/configurations/GcsToGcsConfView';
+import GcsToGcsRunView from '../../widgets/runs/GcsToGcsRunView';
+import store from '@/store';
 import viewJson from '@/components/widgets/parameters/viewJson.vue';
 
 export default {
 	components: {
-		StorageToStorageConfView,
-		StorageToStorageRunView,
+		GcsToGcsConfView,
+		GcsToGcsRunView,
 		viewJson
 	},
 	data: () => ({
@@ -79,10 +71,10 @@ export default {
 		async getRun() {
 			this.$data.isFetchAndAdding = true;
 			//get the conf is not in mirrorExcGcsToGcsRuns
-			if (this.storageToStorageRuns[this.confId] === undefined) {
+			if (this.mirrorExcGcsToGcsRuns[this.confId] === undefined) {
 				await this.getFirestoreData();
 			}
-			this.run = this.storageToStorageRuns[this.runId];
+			this.run = this.mirrorExcGcsToGcsRuns[this.runId];
 			this.$data.isFetchAndAdding = false;
 		},
 		async getFirestoreData() {
@@ -90,10 +82,10 @@ export default {
 			this.$data.fetchAndAddStatus = 'Loading';
 			this.$data.moreToFetchAndAdd = false;
 			try {
-				await store.dispatch('storageToStorageRuns/closeDBChannel', {
+				await store.dispatch('mirrorExcGcsToGcsRuns/closeDBChannel', {
 					clearModule: true
 				});
-				await store.dispatch('storageToStorageRuns/fetchById', runId);
+				await store.dispatch('mirrorExcGcsToGcsRuns/fetchById', runId);
 				this.$data.fetchAndAddStatus = 'Success';
 			} catch (e) {
 				this.$data.fetchAndAddStatus = 'Error';
@@ -106,7 +98,7 @@ export default {
 			isAuthenticated: state => state.user.isAuthenticated,
 			user: state => state.user.user,
 			settings: state => state.settings,
-			storageToStorageRuns: state => state.storageToStorageRuns.data
+			mirrorExcGcsToGcsRuns: state => state.mirrorExcGcsToGcsRuns.data
 		}),
 		runId() {
 			return this.$route.params.runId;
