+++ src/components/data-workflows/gcs-to-gcs/GcsToGcsRuns.vue
@@ -105,11 +105,11 @@
 import { mapState } from 'vuex';
 import { mapGetters } from 'vuex';
 import VueJsonPretty from 'vue-json-pretty';
-import store from '@/store/index';
+import store from '@/store';
 import moment from 'moment';
 import _ from 'lodash';
 import Util from '@/util';
-import DataManagementFilters from './widgets/filters/DataManagementFilters';
+import DataManagementFilters from '../../widgets/filters/DataManagementFilters';
 
 export default {
 	components: {
