+++ src/components/data-workflows/storage-to-table/StorageToTableConf.vue
@@ -3,11 +3,8 @@
 		<v-row v-if="!isFetchAndAdding">
 			<v-col cols="12" offset="0">
 				<v-tabs v-model="activeTab" color="black" background-color="#E0E0E0" slider-color="primary">
-					<v-tab ripple href="#rundetails">
-						Run Details
-					</v-tab>
 					<v-tab ripple href="#confoverview">
-						Configuration
+						Overview
 					</v-tab>
 					<v-tab ripple href="#confschema">
 						Schema
@@ -18,21 +15,12 @@
 					<v-tab ripple href="#conversation">
 						Conversation
 					</v-tab>
-					<v-tab-item value="rundetails">
-						<v-card v-if="this.run !== undefined">
-							<StorageToTableRunView
-								:run="this.run"
-								:runId="this.run.job_id"
-								:activeHeader="true"
-							></StorageToTableRunView
-						></v-card>
-					</v-tab-item>
 					<v-tab-item value="confoverview">
 						<v-card v-if="this.conf !== undefined">
 							<storageToTableConfOverview
 								:conf="this.conf"
 								:confId="this.conf.table_name"
-								:activeHeader="false"
+								:activeHeader="true"
 							></storageToTableConfOverview
 						></v-card>
 					</v-tab-item>
@@ -43,7 +31,7 @@
 					</v-tab-item>
 					<v-tab-item value="fulljson">
 						<v-card>
-							<viewJson :json="this.run" :jsonID="this.run.id" />
+							<viewJson :json="this.conf" :jsonID="this.conf.id" />
 						</v-card>
 					</v-tab-item>
 				</v-tabs>
@@ -59,9 +47,8 @@
 
 <script>
 import { mapState } from 'vuex';
-import store from '@/store/index';
+import store from '@/store';
 import viewJson from '@/components/widgets/parameters/viewJson.vue';
-import StorageToTableRunView from '@/components/widgets/runs/StorageToTableRunView.vue';
 import storageToTableConfOverview from '@/components/widgets/configurations/StorageToTableConfOverview.vue';
 import tableSchemaView from '@/components/widgets/configurations/TableSchemaView.vue';
 
@@ -69,8 +56,7 @@ export default {
 	components: {
 		viewJson,
 		storageToTableConfOverview,
-		tableSchemaView,
-		StorageToTableRunView
+		tableSchemaView
 	},
 	data() {
 		return {
@@ -79,16 +65,18 @@ export default {
 		};
 	},
 	async mounted() {
+		this.isFetchAndAdding = true;
 		await this.getFirestoreData();
 		this.isFetchAndAdding = false;
 	},
 	methods: {
 		async getFirestoreData() {
-			await store.dispatch('storageToTableRun/closeDBChannel', {
+			await store.dispatch('storageToTableConf/closeDBChannel', {
 				clearModule: true
 			});
 			await store
-				.dispatch('storageToTableRun/fetchAndAdd', {
+				.dispatch('storageToTableConf/fetchAndAdd', {
+					sourceId: this.bucketIn,
 					itemId: this.itemId
 				})
 				.catch(console.error);
@@ -96,22 +84,19 @@ export default {
 	},
 	computed: {
 		...mapState({
-			storageToTableRun: state => state.storageToTableRun.data
+			storageToTableConf: state => state.storageToTableConf.data
 		}),
 		itemId() {
-			var itemId = this.$route.params.pathId;
-			return itemId;
+			return this.$route.params.pathId.split('/')[1];
+		},
+		bucketIn() {
+			return this.$route.params.pathId.split('/')[0];
 		},
 		conf() {
-			//Add the bucket file source to the storageToTableRun object
-			var conf = this.storageToTableRun.configuration_context;
-			conf.source_bucket = this.storageToTableRun.source_bucket;
-			conf.id = this.storageToTableRun.filename_template;
+			//Add the bucket file source to the SingleDoc object
+			var conf = this.storageToTableConf;
+			conf.source_bucket = this.bucketIn;
 			return conf;
-		},
-		run() {
-			var run = this.storageToTableRun;
-			return run;
 		}
 	}
 };
