+++ src/components/data-workflows/vm-launcher/VmLauncherConfs.vue
@@ -156,10 +156,10 @@
 import { mapState } from 'vuex';
 import { mapGetters } from 'vuex';
 import VueJsonPretty from 'vue-json-pretty';
-import store from '@/store/index';
+import store from '@/store';
 import _ from 'lodash';
-import DataManagementFilters from './widgets/filters/DataManagementFilters';
-import ActivatedStatusChip from './widgets/datatablewidgets/ActivatedStatusChip.vue';
+import DataManagementFilters from '../../widgets/filters/DataManagementFilters';
+import ActivatedStatusChip from '../../widgets/datatablewidgets/ActivatedStatusChip.vue';
 import ConfsComponent from '@/mixins/confsComponent.js';
 
 export default {
