+++ src/views/GcsToGcsConf.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import GcsToGcsConf from '../components/GcsToGcsConf';
+import GcsToGcsConf from '../components/data-workflows/gcs-to-gcs/GcsToGcsConf';
 import DataManagementHeader from '../components/widgets/layout/DataManagementHeader';
 
 export default {
